+++ test/integration/win/serverspec/serverspec_helper.rb
@@ -1,6 +1,7 @@
 require 'serverspec'
 
 VERSION = '4.0.1'.freeze
+MARIADB_VERSION = '5.1.39'.freeze
 
 if (/cygwin|mswin|mingw|bccwin|wince|emx/ =~ RUBY_PLATFORM).nil?
   set :backend, :exec